Separated authentication and encryption code + moved code in 'utils' into utilities folder.
