+++ lib/did/DidDocument.ts
