+++ lib/did/DidResolver.ts
